move statusbar on level above to shared components
