+++ internal/bubbles/components/statusbar/events.go
