+++ internal/bubbles/components/statusbar/model.go
